# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: (title)
  description: |-
    Calendar Booking API

    Simplified booking system inspired by Cal.com
    Two roles: calendar owner (single static profile) and guests (book without registration)
    All times are in UTC, client handles timezone conversion
  version: 0.0.0
...
paths:
...
  /api/admin/owner/working-hours:
    get:
      operationId: AdminWorkingHours_listWorkingHours
      summary: List working hours schedule (one entry per weekday)
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingHoursListResponse'
      tags:
        - Admin - Owner & Working Hours
    post:
      operationId: AdminWorkingHours_createWorkingHours
      summary: Create working hours entry for a weekday
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingHours'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin - Owner & Working Hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkingHoursRequest'
  /api/admin/owner/working-hours/{weekday}:
    put:
      operationId: AdminWorkingHours_updateWorkingHours
      summary: Update working hours for a specific weekday
      parameters:
        - name: weekday
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DayOfWeek'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingHours'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin - Owner & Working Hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkingHoursRequest'
    delete:
      operationId: AdminWorkingHours_deleteWorkingHours
      summary: Delete working hours for a specific weekday
      parameters:
        - name: weekday
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/DayOfWeek'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin - Owner & Working Hours
...
components:
  schemas:
...
    CreateWorkingHoursRequest:
      type: object
      required:
        - weekday
        - startTime
        - endTime
      properties:
        weekday:
          $ref: '#/components/schemas/DayOfWeek'
        startTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        endTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
      description: Create working hours entry request
    DayOfWeek:
      type: string
      enum:
        - mon
        - tue
        - wed
        - thu
        - fri
        - sat
        - sun
      description: Enum for days of the week
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - SLOT_UNAVAILABLE
            - TIME_IN_PAST
            - OUTSIDE_WORKING_HOURS
            - DATE_RANGE_TOO_LARGE
            - HAS_BOOKINGS
        message:
          type: string
        details:
          type: array
          items:
            type: string
      description: Standard error response
...
    UpdateWorkingHoursRequest:
      type: object
      properties:
        startTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        endTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
      description: Update working hours entry request (partial)
    WorkingHours:
      type: object
      required:
        - id
        - weekday
        - startTime
        - endTime
      properties:
        id:
          type: string
        weekday:
          $ref: '#/components/schemas/DayOfWeek'
        startTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
        endTime:
          type: string
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
      description: Working hours entry for a specific weekday
    WorkingHoursListResponse:
      type: object
      required:
        - workingHours
      properties:
        workingHours:
          type: array
          items:
            $ref: '#/components/schemas/WorkingHours'
      description: Working hours list response
...
servers:
  - url: https://api.example.com
    description: Production server
    variables: {}